requirements.txt
aiomysql==0.2.0
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
kombu==5.4.2
mysqlclient==2.2.4
packaging==24.1
prompt_toolkit==3.0.48
PyJWT==1.7.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
